# apps/bubu-log/openapi.yaml
openapi: 3.1.0
info:
  title: Bubu Log API
  description: Baby activity tracking API
  version: 1.0.0
servers:
  - url: /api
    description: API server
paths:
  /activities:
    get:
      operationId: getActivities
      summary: Get activity list
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/ActivityType'
        - name: types
          in: query
          description: Comma-separated list of activity types to filter
          schema:
            type: string
        - name: date
          in: query
          description: Date in YYYY-MM-DD format (returns activities for this day including cross-midnight)
          schema:
            type: string
            format: date
        - name: startTimeGte
          in: query
          description: Filter activities with startTime >= this value (ISO format)
          schema:
            type: string
            format: date-time
        - name: startTimeLt
          in: query
          description: Filter activities with startTime < this value (ISO format)
          schema:
            type: string
            format: date-time
        - name: crossStartTime
          in: query
          description: When true, also include activities that started before startTimeGte but have endTime >= startTimeGte
          schema:
            type: boolean
            default: false
        - name: crossEndTime
          in: query
          description: When true, also include activities that started before startTimeLt but have endTime >= startTimeLt (i.e., activities that extend past the end boundary)
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Activity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Activity'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      operationId: createActivity
      summary: Create new activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityInput'
      responses:
        '201':
          description: Created activity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '409':
          description: Time conflict with existing activity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityConflictError'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      operationId: batchDeleteActivities
      summary: Batch delete activities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: string
              required:
                - ids
      responses:
        '200':
          description: Deletion success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
        '500':
          $ref: '#/components/responses/ServerError'

    put:
      operationId: batchUpdateActivityDate
      summary: Batch update activity dates
      description: Move selected activities to a different date while preserving time of day
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: string
                  description: Activity IDs to update
                targetDate:
                  type: string
                  format: date
                  description: Target date (YYYY-MM-DD format) to move activities to
              required:
                - ids
                - targetDate
      responses:
        '200':
          description: Update success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
    ActivityType:
      type: string
      enum:
        - SLEEP
        - DIAPER
        - BREASTFEED
        - BOTTLE
        - PUMP
        - HEAD_LIFT
        - PASSIVE_EXERCISE
        - ROLL_OVER
        - PULL_TO_SIT
        - GAS_EXERCISE
        - BATH
        - OUTDOOR
        - EARLY_EDUCATION
        - SUPPLEMENT
        - SPIT_UP
    MilkSource:
      type: string
      enum:
        - BREAST_MILK
        - FORMULA
    PoopColor:
      type: string
      enum:
        - YELLOW
        - GREEN
        - BROWN
        - BLACK
        - WHITE
        - RED
    PeeAmount:
      type: string
      enum:
        - SMALL
        - MEDIUM
        - LARGE
    BreastFirmness:
      type: string
      enum:
        - SOFT    # 软
        - ELASTIC # 弹
        - HARD    # 硬
    SupplementType:
      type: string
      enum:
        - AD    # 维生素AD
        - D3    # 维生素D3
    SpitUpType:
      type: string
      enum:
        - NORMAL     # 普通吐奶
        - PROJECTILE # 喷射性吐奶
...
    Activity:
      type: object
      required:
        - id
        - type
        - startTime
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/ActivityType'
        startTime:
          type: string
          format: date-time
          description: 活动开始时间
        endTime:
          type: string
          format: date-time
          nullable: true
          description: 活动结束时间（正在进行的活动没有结束时间）
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        hasPoop:
          type: boolean
          nullable: true
        hasPee:
          type: boolean
          nullable: true
        poopColor:
          $ref: '#/components/schemas/PoopColor'
        poopPhotoUrl:
          type: string
          nullable: true
        peeAmount:
          $ref: '#/components/schemas/PeeAmount'
        burpSuccess:
          type: boolean
          nullable: true
        milkAmount:
          type: integer
          nullable: true
        milkSource:
          $ref: '#/components/schemas/MilkSource'
          nullable: true
        breastFirmness:
          $ref: '#/components/schemas/BreastFirmness'
          nullable: true
        supplementType:
          $ref: '#/components/schemas/SupplementType'
          nullable: true
        spitUpType:
          $ref: '#/components/schemas/SpitUpType'
          nullable: true
        count:
          type: integer
          nullable: true
          description: 次数（用于翻身、拉坐等计数类活动）
        notes:
          type: string
          nullable: true
    CreateActivityInput:
      type: object
      required:
        - type
        - startTime
      properties:
        type:
          $ref: '#/components/schemas/ActivityType'
        startTime:
          type: string
          format: date-time
          description: 活动开始时间
        endTime:
          type: string
          format: date-time
          description: 活动结束时间（可选）
        hasPoop:
          type: boolean
        hasPee:
          type: boolean
        poopColor:
          $ref: '#/components/schemas/PoopColor'
        poopPhotoUrl:
          type: string
        peeAmount:
          $ref: '#/components/schemas/PeeAmount'
        burpSuccess:
          type: boolean
        milkAmount:
          type: integer
        milkSource:
          $ref: '#/components/schemas/MilkSource'
        breastFirmness:
          $ref: '#/components/schemas/BreastFirmness'
        supplementType:
          $ref: '#/components/schemas/SupplementType'
        spitUpType:
          $ref: '#/components/schemas/SpitUpType'
        count:
          type: integer
          description: 次数（用于翻身、拉坐等计数类活动）
        notes:
          type: string
        force:
          type: boolean
          description: Force create even if there's an overlap warning (not for duplicates)
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    ActivityConflictError:
      type: object
      required:
        - error
        - code
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          enum:
            - DUPLICATE_ACTIVITY
            - OVERLAP_ACTIVITY
          description: |
            DUPLICATE_ACTIVITY - Same type activity exists at same time (cannot be overridden)
            OVERLAP_ACTIVITY - Time overlaps with another activity (can be overridden with force=true)
        conflictingActivity:
          $ref: '#/components/schemas/Activity'
...
  responses:
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'